+++ config/test.exs
@@ -1,0 +1,9 @@
+import Config
+
+config :loxe,
+  on_protocol_undefined: :placeholder,
+  metadata: [
+    except: [
+      :hide_me,
+    ]
+  ]