+++ test/loxe/formatter_test.exs
@@ -5,23 +5,35 @@ defmodule Loxe.FormatterTest do
 
   describe "format/1" do
     test "can handle an empty binary" do
-      assert "msg=\"\"" == Formatter.format("")
+      assert "msg=\"\"" == IO.iodata_to_binary(Formatter.format(""))
     end
 
     test "can handle an binary" do
-      assert "msg=\"Hello, World!\"" == Formatter.format("Hello, World!")
+      assert "msg=\"Hello, World!\"" == IO.iodata_to_binary(Formatter.format("Hello, World!"))
     end
 
     test "can handle a empty list" do
-      assert "msg=\"\"" == Formatter.format([])
+      assert "msg=\"\"" == IO.iodata_to_binary(Formatter.format([]))
     end
 
     test "can handle a iodata list" do
-      assert "msg=\"Hello, World!\"" == Formatter.format(["Hello", [", World", "!"]])
+      assert "msg=\"Hello, World!\"" == IO.iodata_to_binary(Formatter.format(["Hello", [", World", "!"]]))
     end
 
     test "can handle a keyword list" do
-      assert "a=Apple b=Bacon" == Formatter.format([a: "Apple", b: "Bacon"])
+      assert "a=Apple b=Bacon" == IO.iodata_to_binary(Formatter.format([a: "Apple", b: "Bacon"]))
+    end
+
+    test "can handle a unencodable value" do
+      assert "a=Apple b=<protocol-undefined>" == IO.iodata_to_binary(Formatter.format([a: "Apple", b: ~c"Bacon"]))
+    end
+  end
+
+  describe "filter_metadata/1" do
+    test "will filter keys as specified" do
+      assert [
+        something: "Should exist"
+      ] == Formatter.filter_metadata(something: "Should exist", hide_me: "Shouldn't exist")
     end
   end
 end